Profile Fixed

+++ src/main/java/com/example/Repository/Db/EventDbRepo.java
@@ -1,13 +1,8 @@
 package com.example.Repository.Db;
 
 import com.example.Domain.Event;
-
-import com.example.Repository.PagingRepo.Page;
-
-import com.example.Domain.Relationship;
 import com.example.Repository.PagingRepo.Page;
 import com.example.Repository.PagingRepo.PageType;
-
 import com.example.Repository.PagingRepo.Pageble;
 import com.example.Repository.Repository;
 import com.example.Utils.Generator;
@@ -33,21 +28,21 @@ public class EventDbRepo extends DbRepoId<Long,Event> implements Repository<Long
     @Override
     public boolean save(Event entity) {
         if(entity.getId()==null) entity.setId(generateId());
-        super.sql= "insert into public.\"Events\" values (?, ?, ?,?)";
+        if(super.sql==null)super.sql= "insert into public.\"Events\" values (?, ?, ?,?)";
         return super.save(entity);
     }
 
     @Override
     public Event get(Long id) {
-        super.sql= "select * from public.\"Events\" where id="+id.toString();
+        if(super.sql==null)super.sql= "select * from public.\"Events\" where id="+id.toString();
         return super.get(id);
     }
 
     @Override
     public boolean update(Long id, Event entity) {
-        // if(get(id)==null) throw new RelationshipRepoException("There is no event with that id");
+        //if(get(id)==null) throw new RelationshipRepoException("There is no event with that id");
 
-        super.sql= "update public.\"Events\" set id=?,name=?,the_date=?, description=? where id=?";
+        if(super.sql==null)super.sql= "update public.\"Events\" set id=?,name=?,the_date=?, description=? where id=?";
 
         return super.update(id, entity);
 
@@ -56,38 +51,38 @@ public class EventDbRepo extends DbRepoId<Long,Event> implements Repository<Long
     @Override
     public boolean delete(Long id) {
         //if(get(id)==null) throw new RelationshipRepoException("There is no event with that id");
-        super.sql="delete from public.\"Events\" where id=?";
+        if(super.sql==null)super.sql="delete from public.\"Events\" where id=?";
         return super.delete(id);
     }
 
     @Override
     protected void deleteAll() {
-        super.sql= "delete from public.\"Events\" where id != 0";
+        if(super.sql==null)super.sql= "delete from public.\"Events\" where id != 0";
         super.deleteAll();
     }
 
     @Override
     public int getSize() {
-        super.sql="select count(*) as \"size\" from public.\"Events\"";
+        if(super.sql==null)super.sql="select count(*) as \"size\" from public.\"Events\"";
         return super.getSize();
     }
 
     @Override
     public List<Event> getAll() {
-        super.sql="select * from public.\"Events\"";
+        if(super.sql==null)super.sql="select * from public.\"Events\"";
         return super.getAll();
     }
 
     @Override
     public Page<Event> getAll(Pageble pageble) {
-        super.sql="select * from ( select * ,ROW_NUMBER() over (order by id ASC) as rowss from public.\"Events\" E inner join public.\"UserEvents\" UE on U.id=UE.id_e)as Foo where rowss>=? and rowss<? ";
+        if(super.sql==null)super.sql="select * from ( select * ,ROW_NUMBER() over (order by id_ue ASC) as rowss from public.\"Events\" E inner join public.\"UserEvents\" UE on E.id=UE.id_e)as Foo where rowss>=? and rowss<? ";
         return super.getAll(pageble);
         //return super.getAll();
     }
 
 
     public Page<Event> getPageEvents(PageType type) {
-        super.sql="select * from ( select * ,ROW_NUMBER() over (order by id_rel ASC) as rowss from (select * from public.\"Events\"  as Foo where rowss>=? and rowss<? ";
+        if(super.sql==null)super.sql="select * from ( select * ,ROW_NUMBER() over (order by id_ue ASC) as rowss from public.\"Events\" ) as Foo where rowss>=? and rowss<? ";
         switch(type){
             case CURRENT -> {return super.getCurrentPage();}
             case NEXT -> {return super.getNextPage();}
@@ -97,11 +92,16 @@ public class EventDbRepo extends DbRepoId<Long,Event> implements Repository<Long
         //return super.getAll();
     }
 
+    public List<Event> getUserEvents(Long id){
+        if(super.sql==null)super.sql="select * from public.\"Events\" E inner join public.\"UserEvents\" UE on E.id=UE.id_e where UE.id_u="+id.toString();
+        return super.getAll();
+    }
+
 
 
     @Override
     public List<Long> getAllIds() {
-        super.sql= "select id from public.\"Events\"";
+        if(super.sql==null)super.sql= "select id from public.\"Events\"";
         return super.getAllIds();
     }
